+++ src/components/modules/blockManager.ts
@@ -552,12 +552,37 @@ export default class BlockManager extends Module {
    * Swap Blocks Position
    * @param {Number} fromIndex
    * @param {Number} toIndex
+   * @deprecated — use 'move' instead
    */
   public swap(fromIndex, toIndex): void {
     /** Move up current Block */
     this._blocks.swap(fromIndex, toIndex);
 
     /** Now actual block moved up so that current block index decreased */
+    this.currentBlockIndex = toIndex;
+  }
+
+  /**
+   * Move a block to a new index
+   * @param {Number} toIndex
+   * @param {Number} fromIndex
+   */
+  public move(toIndex, fromIndex = this.currentBlockIndex): void {
+    // make sure indexes are valid and within a valid range
+    if (isNaN(toIndex) || isNaN(fromIndex)) {
+      _.log(`Warning during 'move' call: incorrect indices provided.`, 'warn');
+      return;
+    }
+
+    if (!this.validateIndex(toIndex) || !this.validateIndex(fromIndex)) {
+      _.log(`Warning during 'move' call: indices cannot be lower than 0 or greater than the amount of blocks.`, 'warn');
+      return;
+    }
+
+    /** Move up current Block */
+    this._blocks.move(toIndex, fromIndex);
+
+    /** Now actual block moved so that current block index changed */
     this.currentBlockIndex = toIndex;
   }
 
@@ -603,5 +628,17 @@ export default class BlockManager extends Module {
     Listeners.on(block.holder, 'keyup', (event) => BlockEvents.keyup(event));
     Listeners.on(block.holder, 'dragover', (event) => BlockEvents.dragOver(event as DragEvent));
     Listeners.on(block.holder, 'dragleave', (event) => BlockEvents.dragLeave(event as DragEvent));
+  }
+
+  /**
+   * Validates that the given index is not lower than 0 or higher than the amount of blocks
+   * @param {number} index - index of blocks array to validate
+   */
+  private validateIndex(index: number): boolean {
+    if (index < 0 || index >= this._blocks.length) {
+      return false;
+    }
+
+    return true;
   }
 }